Add files via upload
These files are examples taken from the library database system. They consist of a Java Class that handles the registration and modification of user information, as well as a Java Form that is used in the user interface. These files are essential components of the library database system, allowing users to register as members and make changes to their personal details. The Java Class provides the necessary functionality to process user information, while the Java Form enhances the user experience by providing a visually appealing and interactive interface. These files play a crucial role in the overall functioning of the library database system, ensuring smooth user registration and seamless user management.